Move from workshop/Java/workshop14 to workshop/workshop14
